+++ tests/h2integrate/test_all_examples.py
@@ -4,6 +4,7 @@ import importlib
 from pathlib import Path
 
 import pytest
+import openmdao.api as om
 
 from h2integrate.core.h2integrate_model import H2IntegrateModel
 
@@ -24,20 +25,27 @@ def test_steel_example(subtests):
     model.post_process()
     # Subtests for checking specific values
     with subtests.test("Check LCOH"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOH"), rel=1e-3) == 7.47944016
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOH"), rel=1e-3)
+            == 7.47944016
+        )
 
     with subtests.test("Check LCOS"):
         assert pytest.approx(model.prob.get_val("steel.LCOS"), rel=1e-3) == 1213.87728644
 
     with subtests.test("Check total adjusted CapEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_capex_adjusted"), rel=1e-3)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_capex_adjusted"), rel=1e-3
+            )
             == 5.10869916e09
         )
 
     with subtests.test("Check total adjusted OpEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_opex_adjusted"), rel=1e-3)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_opex_adjusted"), rel=1e-3
+            )
             == 96349901.77625626
         )
 
@@ -93,22 +101,32 @@ def test_simple_ammonia_example(subtests):
 
     with subtests.test("Check total adjusted CapEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_capex_adjusted"), rel=1e-3)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_capex_adjusted"), rel=1e-3
+            )
             == 2.76180599e09
         )
 
     with subtests.test("Check total adjusted OpEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_opex_adjusted"), rel=1e-3)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_opex_adjusted"), rel=1e-3
+            )
             == 66599592.71371833
         )
 
     # Currently underestimated compared to the Reference Design Doc
     with subtests.test("Check LCOH"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOH"), rel=1e-3) == 4.39187968
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOH"), rel=1e-3)
+            == 4.39187968
+        )
     # Currently underestimated compared to the Reference Design Doc
     with subtests.test("Check LCOA"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOA"), rel=1e-3) == 1.06313924
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOA"), rel=1e-3)
+            == 1.06313924
+        )
 
 
 def test_ammonia_synloop_example(subtests):
@@ -156,26 +174,46 @@ def test_ammonia_synloop_example(subtests):
 
     with subtests.test("Check total adjusted CapEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_capex_adjusted"), rel=1e-6)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_capex_adjusted"), rel=1e-6
+            )
             == 3.83856529e09
         )
 
     with subtests.test("Check total adjusted OpEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_opex_adjusted"), rel=1e-6)
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_opex_adjusted"), rel=1e-6
+            )
             == 81093533.8566508
         )
 
     with subtests.test("Check LCOH"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOH"), rel=1e-6) == 5.85374921
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOH"), rel=1e-6)
+            == 5.85374921
+        )
 
     with subtests.test("Check LCOA"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOA"), rel=1e-6) == 1.10368921
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOA"), rel=1e-6)
+            == 1.10368921
+        )
 
 
 def test_wind_h2_opt_example(subtests):
     # Change the current working directory to the example's directory
     os.chdir(examples_dir / "05_wind_h2_opt")
+
+    # Run without optimization
+    model_init = H2IntegrateModel(Path.cwd() / "wind_plant_electrolyzer0.yaml")
+
+    # Run the model
+    model_init.run()
+
+    model_init.post_process()
+
+    annual_h20 = model_init.prob.get_val("electrolyzer.total_hydrogen_produced", units="kg/year")[0]
 
     # Create a H2Integrate model
     model = H2IntegrateModel(Path.cwd() / "wind_plant_electrolyzer.yaml")
@@ -185,27 +223,67 @@ def test_wind_h2_opt_example(subtests):
 
     model.post_process()
 
-    with subtests.test("Check LCOH"):
-        assert model.prob.get_val("financials_group_1.LCOH")[0] < 4.64
+    with subtests.test("Check initial H2 production"):
+        assert annual_h20 < (60500000 - 10000)
 
     with subtests.test("Check LCOE"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOE"), rel=1e-3) == 0.09009908
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOE")[0], rel=1e-3)
+            == 0.151189
+        )
+
+    with subtests.test("Check electrolyzer size"):
+        assert (
+            pytest.approx(model.prob.get_val("electrolyzer.electrolyzer_size_mw")[0], rel=1e-3)
+            == 1500.0
+        )
+    # Read the resulting SQL file and compare initial and final LCOH values
+
+    sql_path = None
+    for root, _dirs, files in os.walk(Path.cwd()):
+        for file in files:
+            if file == "wind_h2_opt.sql":
+                sql_path = Path(root) / file
+                break
+        if sql_path:
+            break
+    assert (
+        sql_path is not None
+    ), "wind_h2_opt.sql file not found in current working directory or subdirectories."
+
+    cr = om.CaseReader(str(sql_path))
+    cases = list(cr.get_cases())
+    assert len(cases) > 1, "Not enough cases recorded in SQL file."
+
+    # Get initial and final LCOH values
+    initial_lcoh = cases[0].outputs["financials_group_default.LCOH"][0]
+    final_lcoh = cases[-1].outputs["financials_group_default.LCOH"][0]
+
+    with subtests.test("Check LCOH changed"):
+        assert final_lcoh != initial_lcoh
 
     with subtests.test("Check total adjusted CapEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_capex_adjusted"), rel=1e-3)
-            == 1.82152792e09
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_capex_adjusted")[0], rel=1e-3
+            )
+            == 2783126102
         )
-
     with subtests.test("Check total adjusted OpEx"):
         assert (
-            pytest.approx(model.prob.get_val("financials_group_1.total_opex_adjusted"), rel=1e-3)
-            == 51995875.99756081
+            pytest.approx(
+                model.prob.get_val("financials_group_default.total_opex_adjusted")[0], rel=1e-3
+            )
+            == 75543899
         )
 
     with subtests.test("Check minimum total hydrogen produced"):
         assert (
-            model.prob.get_val("electrolyzer.total_hydrogen_produced", units="kg/year") >= 60500000
+            pytest.approx(
+                model.prob.get_val("electrolyzer.total_hydrogen_produced", units="kg/year")[0],
+                abs=10000,
+            )
+            == 60500000
         )
 
 
@@ -246,10 +324,16 @@ def test_wind_wave_doc_example(subtests):
 
     # Subtests for checking specific values
     with subtests.test("Check LCOC"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOC"), rel=1e-3) == 2.26955589
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOC"), rel=1e-3)
+            == 2.26955589
+        )
 
     with subtests.test("Check LCOE"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOE"), rel=1e-3) == 1.05281478
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOE"), rel=1e-3)
+            == 1.05281478
+        )
 
 
 def test_hydro_example(subtests):
@@ -264,11 +348,14 @@ def test_hydro_example(subtests):
 
     model.post_process()
 
-    print(model.prob.get_val("financials_group_1.LCOE"))
+    print(model.prob.get_val("financials_group_default.LCOE"))
 
     # Subtests for checking specific values
     with subtests.test("Check LCOE"):
-        assert pytest.approx(model.prob.get_val("financials_group_1.LCOE"), rel=1e-3) == 0.17653979
+        assert (
+            pytest.approx(model.prob.get_val("financials_group_default.LCOE"), rel=1e-3)
+            == 0.17653979
+        )
 
 
 def test_hybrid_energy_plant_example(subtests):
@@ -287,8 +374,61 @@ def test_hybrid_energy_plant_example(subtests):
     with subtests.test("Check LCOE"):
         assert (
             pytest.approx(
-                model.prob.get_val("financials_group_1.LCOE", units="USD/MW/h")[0],
+                model.prob.get_val("financials_group_default.LCOE", units="USD/MW/h")[0],
                 rel=1e-5,
             )
             == 83.2123
+        )
+
+
+def test_asu_example(subtests):
+    # Change the current working directory to the example's directory
+    os.chdir(examples_dir / "13_air_separator")
+
+    # Create a H2Integrate model
+    model = H2IntegrateModel(Path.cwd() / "13_air_separator.yaml")
+
+    # Run the model
+    model.run()
+
+    model.post_process()
+
+    # Subtests for checking specific values
+    with subtests.test("Check LCON"):
+        assert (
+            pytest.approx(
+                model.prob.get_val("financials_group_default.LCON", units="USD/kg")[0],
+                abs=1e-4,
+            )
+            == 0.309041977334972
+        )
+
+
+def test_hydrogen_dispatch_example(subtests):
+    # Change the current working directory to the example's directory
+    os.chdir(examples_dir / "14_wind_hydrogen_dispatch")
+
+    # Create a H2Integrate model
+    model = H2IntegrateModel(Path.cwd() / "inputs" / "h2i_wind_to_h2_storage.yaml")
+
+    model.run()
+
+    model.post_process()
+
+    with subtests.test("Check LCOE"):
+        assert (
+            pytest.approx(
+                model.prob.get_val("financials_group_default.LCOE", units="USD/MW/h")[0],
+                rel=1e-5,
+            )
+            == 106.13987
+        )
+
+    with subtests.test("Check LCOH"):
+        assert (
+            pytest.approx(
+                model.prob.get_val("financials_group_default.LCOH", units="USD/kg")[0],
+                rel=1e-5,
+            )
+            == 5.68452215
         )