+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## 0.3.1, Unreleased, TBD
+- Added rule-based storage system control capability
+- Bugfix: only run pyxdsm when there are connections in the system
 - Added *_out/ to .gitignore to avoid clutter
 - Added methanol production base class
 - Added steam methane reforming methanol production technology
@@ -29,6 +31,14 @@
 - Added ammonia synloop, partially addressing https://github.com/NREL/H2Integrate/issues/169
 - Added geologic hydrogen (geoh2) converter and examples [PR 135](https://github.com/NREL/H2Integrate/pull/135)
 - Changed plant_config `atb_year` entry to `financial_analysis_start_year`
+- Allow users to specify the technologies to include in the metrics calculations in the plant configuration file
+- Update PySAM solar performance model to allow for more user-configurability.
+- Added simple air separation unit (ASU) converter to model nitrogen production.
+- Added two ATB-compatible solar-PV cost models.
+- Changed electrolyzer capacity to be specified as `n_clusters` rather than `rating` in electrolyzer performance model config
+- Changed electrolyzer capacity to be an input to the electrolyzer cost models rather than pulled from the cost model config.
+- Updated finance parameter organization naming in `plant_config`.
+
 
 ## 0.3.0 [May 2 2025]